Add hicstraw adapter for Juicer 2.0+ compatibility
