+++ README.md
@@ -19,6 +19,7 @@ To this end, we developed Quagga, a tool for detection and statistical verificat
   - six
   - h5py
   - requests
+  - hic-straw
   
   
   **Install from GitHub**
@@ -98,6 +99,18 @@ To this end, we developed Quagga, a tool for detection and statistical verificat
   - Micro-C: nstrata_blank=10, norm="balanced", threshold=0.15, max_range=5000000, resolution=5000, min_length=500000, min_distance=500000, max_width=50000, window_size=20, sigma=6, rel_height=0.1
 
 [*] If you wish to use negative binomial statistics instead of Poisson statistics, include one additional "nbinom_p=0.8" to specify the p value (success probability) of negative binomial distribution. Quagga will automatically switch to negative binomial mode.
+
+#### Update
+
+**`.hic` File Compatibility**:
+
+Quagga uses the modern `hic-straw` library to read .hic files, which supports all file format versions including those generated by Juicer 2.0+. Previously, `hic_straw.py` only supported version 1.6, and failed with a `TypeError: 'int' object is not subscriptable`.
+
+Your .hic files must contain pre-computed normalization vectors (KR, VC, etc.) for balanced normalization to work. If your files lack these vectors, add them:
+
+```bash
+java -jar juicer_tools.jar addNorm -k KR file.hic
+```
 
 ### Output
